refactor: add thumbnail / refactor code for CR
